mt5-dashboard-api/requirements.txt
# FastAPI Core
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Templates & Static Files
jinja2>=3.1.0
python-multipart>=0.0.9

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2

# Visualization
plotly>=5.17.0

# Configuration
pyyaml>=6.0

# Caching (for performance optimization)
cachetools>=5.3.0

# Monitoring (optional but recommended for VPS)
psutil>=5.9.0
